[DEV-3737] Revert most fixes to apply_annotations_to_sql since it's being reworked
