+++ usaspending_api/download/filestreaming/__init__.py
@@ -1,0 +1,1 @@
+NAMING_CONFLICT_SUFFIX = "_NAMING_CONFLICT_SUFFIX"
